Fixed id validations in all methods.

+++ src/controllers/employees.js
@@ -50,36 +50,43 @@ const getAllEmployees = async (req, res) => {
 };
 
 const updateEmployee = async (req, res) => {
-  try {
-    if (!req.params) {
-      res.status(400).json({
-        message: 'Missing id parameter',
+  if (req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+    try {
+      if (!req.params) {
+        res.status(400).json({
+          message: 'Missing id parameter',
+          data: undefined,
+          error: true,
+        });
+      }
+      const result = await Employee.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+      });
+      if (!result) {
+        return res.status(404).json({
+          message: 'Employee not found',
+          data: undefined,
+          error: true,
+        });
+      }
+      return res.status(201).json({
+        message: 'Employee information updated',
+        data: result,
+        error: false,
+      });
+    } catch (error) {
+      return res.json({
+        message: 'There has been an error',
         data: undefined,
         error: true,
       });
     }
-    const result = await Employee.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-    if (!result) {
-      return res.status(404).json({
-        message: 'Employee not found',
-        data: undefined,
-        error: true,
-      });
-    }
-    return res.status(201).json({
-      message: 'Employee information updated',
-      data: result,
-      error: false,
-    });
-  } catch (error) {
-    return res.json({
-      message: 'There has been an error',
-      data: undefined,
-      error: true,
-    });
   }
+  return res.status(404).json({
+    message: 'Invalid ID',
+    data: undefined,
+    error: true,
+  });
 };
 
 const getEmployeesById = async (req, res) => {
@@ -114,27 +121,34 @@ const getEmployeesById = async (req, res) => {
 };
 
 const deleteEmployee = async (req, res) => {
-  try {
-    const result = await Employee.findByIdAndDelete(req.params.id);
-    if (!result) {
+  if (req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+    try {
+      const result = await Employee.findByIdAndDelete(req.params.id);
+      if (!result) {
+        return res.status(404).json({
+          message: 'Employee not found',
+          data: undefined,
+          error: true,
+        });
+      }
+      return res.status(200).json({
+        message: 'The employee has been deleted',
+        data: result,
+        error: false,
+      });
+    } catch (error) {
       return res.status(404).json({
-        message: 'Employee not found',
+        message: error.message,
         data: undefined,
         error: true,
       });
     }
-    return res.status(204).json({
-      message: 'The employee has been deleted',
-      data: result,
-      error: false,
-    });
-  } catch (error) {
-    return res.status(404).json({
-      message: error.message,
-      data: undefined,
-      error: true,
-    });
   }
+  return res.status(404).json({
+    message: 'Invalid ID',
+    data: undefined,
+    error: true,
+  });
 };
 
 export default {
